Move InstantLongSerializer to Indispensable and deprecate
