+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/io/InstantLongSerializer.kt
@@ -1,4 +1,4 @@
-package at.asitplus.signum.indispensable.cosef
+package at.asitplus.signum.indispensable.io
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
@@ -8,7 +8,13 @@ import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import kotlin.time.Instant
 
-@Deprecated("Moved", replaceWith = ReplaceWith("at.asitplus.signum.indispensable.io.InstantLongSerializer" ))
+/**
+ * Serializes [Instant] as whole seconds from the Unix epoch.
+ *
+ * In JSON this is a NumericDate encoded as a JSON number (RFC 7519, Section 2).
+ * In CBOR this is a CWT NumericDate encoded as an untagged CBOR numeric date, with
+ * CBOR tag 1 omitted (RFC 8392, Section 2; RFC 8949, Section 3.4.2).
+ */
 object InstantLongSerializer : KSerializer<Instant> {
 
     override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("InstantLongSerializer", PrimitiveKind.LONG)
